refactor(core-rs): split skills.rs into skills/{scan,crud,agents}.rs (todo #317)
skills.rs was 604 lines, over the 300-line file cap. Splits it into a facade
plus scan/crud/agents modules (Decision D1) and wires list_agents/
create_agent/update_agent through crate::agent_description::
derive_agent_description instead of the old first-line heuristic, so a
frontmatter-declared description is used when present.
